refactor: move fedkseed to homo  (#47)
